docs: use single premium-submitted page with tier query param
